[Commons] Change some db-related fields type

+++ endpoints/middleware.go
@@ -13,8 +13,8 @@ type RequestWithID struct {
 }
 
 type ListRequest struct {
-	Limit  int64
-	Offset int64
+	Limit  int
+	Offset int
 }
 
 // LoggingMiddleware returns an endpoint middleware that logs the
@@ -23,12 +23,10 @@ type ListRequest struct {
 func LoggingMiddleware(logger log.Logger) endpoint.Middleware {
 	return func(next endpoint.Endpoint) endpoint.Endpoint {
 		return func(ctx context.Context, request interface{}) (response interface{}, err error) {
-
 			defer func(begin time.Time) {
 				logger.Log("transport_error", err, "took", time.Since(begin))
 			}(time.Now())
 			return next(ctx, request)
-
 		}
 	}
 }
